requirements.txt
Flask==3.0.0
Werkzeug==3.0.1
Pillow==10.1.0
python-dotenv==1.0.0
python-dateutil==2.8.2
Flask-Mail==0.9.1
gunicorn==21.2.0